BHBC-1937: Make wfs-utils file
